+++ packages/better-effect-better-auth/tests/package-foundation.test.ts
@@ -1,6 +1,7 @@
 import { expect, test } from 'bun:test'
 
 import * as publicApi from '../src'
+import * as honoApi from '../src/hono'
 
 test('publishes only the implemented runtime API', () => {
   expect(Object.keys(publicApi).sort()).toEqual([
@@ -8,4 +9,5 @@ test('publishes only the implemented runtime API', () => {
     'BetterAuthApiError',
     'Unauthenticated'
   ])
+  expect(Object.keys(honoApi).sort()).toEqual(['BetterAuthHono'])
 })